LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.3_Future_Reserved_Words/S7.5.3_A1.7-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token '='. Debugger keyword must be followed by a ';'.
S7.5.3_A1.7

PASS Expected parsing failure

TEST COMPLETE
